rename old plans tab and start work on new plans tab
