+++ src/components/Nav/Nav.tsx
@@ -6,13 +6,13 @@ import Tab_Events from "./Tabs/Tab_Events/Tab_Events";
 import Tab_Insights from "./Tabs/Tab_Insights/Tab_Insights";
 import Tab_Optimizer from "./Tabs/Tab_Optimizer/Tab_Optimizer";
 import Tab_Plans from "./Tabs/Tab_Plans/Tab_Plans";
+import Tab_Planz from "./Tabs/Tab_Plans/_Tab_Plans";
 import Tab_Settings from "./Tabs/Tab_Settings/Tab_Settings";
 import classes from "./Nav.module.css";
 import Search from "./Search/Search";
 import { createRef, useState } from "react";
 import { useMediaQuery } from "@mantine/hooks";
 import FiltersDrawer from "./FiltersDrawer/FiltersDrawer";
-import Tab_Sections from "./Tabs/Tab_Sections/Tab_Sections";
 
 const tabData = [
   //this array contains the data for each tab
@@ -25,10 +25,10 @@ const tabData = [
     component: Tab_Plans,
   },
   {
-    value: "sections",
-    label: "sections",
-    icon: "view_timeline",
-    component: Tab_Sections,
+    value: "planz",
+    label: "planz",
+    icon: "list",
+    component: Tab_Planz,
   },
   {
     value: "settings",
